Repairs Date.addMonths, adds test case to cover this bug
The method Date.addMonths fails around the year 0. For example, substracting
one month from 0-1-1 leads to 0-12-1 instead of -1-12-1. This is due to tricky
numbering of the months when considering negative years.

A test case is added with does fuzz testing to see if Date.addMonths is
revertable. To actually be revertable, the day is restricted to 1. It fails for
the old implementation, but passes with the new method.